requirements.txt
flask>=2.0.0
flask-socketio>=5.0.0
mysql-connector-python>=8.0.0
smartapi-python>=1.0.0
pyotp>=2.6.0
logzero>=1.7.0